requirements.txt
anyio==4.3.0
APScheduler==3.6.3
certifi==2024.2.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
mysql-connector-python==8.3.0
mysqlclient==2.2.4
python-dotenv==1.0.1
python-telegram-bot==13.5
pytz==2024.1
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.0
tornado==6.4
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.15
